[FEAT] : Layout에서 메뉴처리하던 부분을 Page.tsx 단위로 이동 시킨 후 use Client로 변경 KAN-377
